bring back register-shim.js for browser
