+++ package.json
@@ -3,6 +3,9 @@
   "version": "1.1.0",
   "description": "Resolve any installed ES6 compatible promise",
   "main": "index.js",
+  "browser": {
+    "./register.js": "./register-shim.js"
+  },
   "scripts": {
     "test": "make test"
   },
