+++ register-shim.js
@@ -1,0 +1,2 @@
+"use strict";
+module.exports = require('./lib')(true);
